made Smart class
